[fix] windows Failed to seekTo longer than 3:30s
by @maintel

+++ packages/audioplayers_windows/windows/audioplayers_windows_plugin.cpp
@@ -156,7 +156,9 @@ void AudioplayersWindowsPlugin::HandleMethodCall(
     result->Success(EncodableValue(1));
   } else if (method_call.method_name().compare("seek") == 0) {
     auto position = GetArgument<int>("position", args, (int)(player->GetPosition() / 10000));
-    player->SeekTo(position * 10000);
+    // Thanks to @maintel
+    player->SeekTo(static_cast<int64_t>(position * 10000.0));
+    // endThanks
     result->Success(EncodableValue(1));
   } else if (method_call.method_name().compare("setSourceUrl") == 0) {
     auto url = GetArgument<std::string>("url", args, std::string());
